pyproject.toml
[tool.poetry]
name = "app_in_the_air"
version = "0.1.0"
description = ""
authors = ["Nikita Davydov"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.2"
celery = "^4.4.0"
google-cloud-storage = "^1.25.0"
google-cloud-ndb = "^0.2.2"
google-cloud-datastore = "^1.10.0"
google_auth_oauthlib = "^0.4.1"
google-api-python-client = "^1.7.11"
python-dotenv = "^0.10.5"
django-rest-framework = "^0.1.0"
grpcio = "^1.26.0"
oauth2client = "^4.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
